refactor midi and config functionality into hooks
